Split the classes into separate files. Small fixes. All local tests run.
